Added helper classes for MqttAASRegistry, MqttAASAggregator, MqttSubmodelAPI
