+++ src/main/java/org/eclipse/basyx/extensions/aas/aggregator/mqtt/MqttAASAggregatorHelper.java
@@ -1,0 +1,5 @@
+package org.eclipse.basyx.extensions.aas.aggregator.mqtt;
+
+public class MqttAASAggregatorHelper {
+
+}
