+++ src/main/java/org/eclipse/basyx/extensions/submodel/mqtt/MqttSubmodelAPIHelper.java
@@ -1,0 +1,5 @@
+package org.eclipse.basyx.extensions.submodel.mqtt;
+
+public class MqttSubmodelAPIHelper {
+
+}
